chore: create landing header block
