+++ packages/client/src/pages/landing/components/header/index.ts
@@ -1,0 +1,1 @@
+export { LandingHeader } from './view'
