+++ packages/client/src/pages/landing/components/index.ts
@@ -2,3 +2,4 @@ export { LandingAbout } from './about'
 export { LandingBanner } from './banner'
 export { LandingFooter } from './footer'
 export { LandingGameFlow } from './gameflow'
+export { LandingHeader } from './header'
